# splinterd/api/static/openapi.yml
...
openapi: '3.0.0'
info:
  version: 0.3.12
  title: SplinterD API
  description: REST API for SplinterD
servers:
  - url: http://localhost:9000/api
paths:
...
  /admin/keys:
    get:
      tags:
        - Key Registry
      description: List public key information in the Key Registry
      parameters:
        - $ref: "#/components/parameters/protocol_version"
        - name: offset
          in: query
          description: paging offset
          required: false
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          description: maximum number of items to return (max 100)
          required: false
          schema:
            type: integer
            default: 100
      responses:
        200:
          description: list of public key data objects
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PublicKeyInfo'
                  paging:
                    $ref: '#/components/schemas/Paging'
        400:
          description: Request was malformed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /admin/keys/{public_key}:
    get:
      tags:
        - Key Registry
      description: Fetch public key information in the Key Registry by public key
      parameters:
        - $ref: "#/components/parameters/protocol_version"
        - name: public_key
          in: path
          description: public key to query, in hex
          required: true
          schema:
            type: string
      responses:
        200:
          description: Node
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/PublicKeyInfo"
        400:
          description: "{public_key} was malformed"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: The information for {public_key} was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  parameters:
    protocol_version:
      schema:
        type: integer
        example: 1
      description: The protocol version which the client can understand
      in: header
      name: SplinterProtocolVersion
  schemas:
    Error:
      additionalProperties: false
      properties:
        message:
          description: A message describing the error that occurred
          type: string
          example: DatabaseError({description})
      required:
        - message
...
    PublicKeyInfo:
      type: object
      properties:
        public_key:
          type: string
        node_id:
          type: string
        metadata:
          type: object
          additionalProperties: true
          example:
            name: Jane User
            organization: Acme Corporation
...
    Paging:
      type: object
      properties:
        current:
          type: string
        offset:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        first:
          type: string
        prev:
          type: string
        next:
          type: string
        last:
          type: string
      example:
        current: /admin/nodes?offset=10&limit=10
        offset: 10
        limit: 10
        total: 50
        first: /admin/nodes?offset=0&limit=10
        prev: /admin/nodes?offset=0&limit=10
        next: /admin/nodes?offset=20&limit=10
        last: /admin/nodes?offset=40&limit=10
...